Patch 0.9.1 -Refactoring
