+++ src/main/java/me/shadowalzazel/mcodyssey/alchemy/AlchemyPotions.kt
@@ -1,6 +1,6 @@
 package me.shadowalzazel.mcodyssey.alchemy
 
-import me.shadowalzazel.mcodyssey.alchemy.potions.*
+import me.shadowalzazel.mcodyssey.items.arcane.*
 import me.shadowalzazel.mcodyssey.items.utility.OdysseyPotion
 
 object AlchemyPotions {
