+++ src/main/java/me/shadowalzazel/mcodyssey/items/arcane/rune_items.kt
@@ -1,0 +1,2 @@
+package me.shadowalzazel.mcodyssey.items.arcane
+
