--- src/main/java/me/shadowalzazel/mcodyssey/items/weapons/exoticWeapons.kt
@@ -1,8 +1,0 @@
-package me.shadowalzazel.mcodyssey.items.weapons
-
-import me.shadowalzazel.mcodyssey.constants.OdysseyItemModels
-import me.shadowalzazel.mcodyssey.items.utility.OdysseyWeapon
-import me.shadowalzazel.mcodyssey.items.utility.WeaponTypes
-import org.bukkit.Material
-
-object AbzuBlade: OdysseyWeapon("Abzu Blade", Material.NETHERITE_SWORD, OdysseyItemModels.ABZU_BLADE, 18.75, 0.7, WeaponTypes.SWORD)
