+++ src/main/java/me/shadowalzazel/mcodyssey/items/weapons/pole_arms.kt
@@ -1,0 +1,18 @@
+package me.shadowalzazel.mcodyssey.items.weapons
+
+import me.shadowalzazel.mcodyssey.constants.OdysseyItemModels
+import me.shadowalzazel.mcodyssey.items.utility.OdysseyWeapon
+import me.shadowalzazel.mcodyssey.items.utility.WeaponTypes
+import org.bukkit.Material
+
+/* --------------------------WARHAMMER---------------------------*/
+
+object WoodenWarhammer: OdysseyWeapon("Wooden Warhammer", Material.WOODEN_PICKAXE, OdysseyItemModels.WOODEN_WARHAMMER, 6.5, 0.7, WeaponTypes.POLE_ARM)
+
+object GoldenWarhammer: OdysseyWeapon("Golden Warhammer", Material.GOLDEN_PICKAXE, OdysseyItemModels.GOLDEN_WARHAMMER, 7.5, 0.7, WeaponTypes.POLE_ARM)
+
+object StoneWarhammer: OdysseyWeapon("Stone Warhammer", Material.STONE_PICKAXE, OdysseyItemModels.STONE_WARHAMMER, 8.5, 0.7, WeaponTypes.POLE_ARM)
+
+object IronWarhammer: OdysseyWeapon("Iron Warhammer", Material.IRON_PICKAXE, OdysseyItemModels.IRON_WARHAMMER, 9.5, 0.7, WeaponTypes.POLE_ARM)
+
+object DiamondWarhammer: OdysseyWeapon("Diamond Warhammer", Material.DIAMOND_PICKAXE, OdysseyItemModels.DIAMOND_WARHAMMER, 10.5, 0.7, WeaponTypes.POLE_ARM)
