+++ src/main/java/me/shadowalzazel/mcodyssey/items/weapons/spears.kt
@@ -7,27 +7,27 @@ import org.bukkit.Material
 
 /* --------------------------SPEAR---------------------------*/
 
-object WoodenSpear: OdysseyWeapon("Wooden Spear", Material.WOODEN_SHOVEL, OdysseyItemModels.WOODEN_SPEAR, 2.5, 0.85, WeaponTypes.POLE_ARM)
+object WoodenSpear: OdysseyWeapon("Wooden Spear", Material.WOODEN_SHOVEL, OdysseyItemModels.WOODEN_SPEAR, 2.5, 0.85, WeaponTypes.SPEAR)
 
-object GoldenSpear: OdysseyWeapon("Golden Spear", Material.GOLDEN_SHOVEL, OdysseyItemModels.GOLDEN_SPEAR, 2.5, 0.85, WeaponTypes.POLE_ARM)
+object GoldenSpear: OdysseyWeapon("Golden Spear", Material.GOLDEN_SHOVEL, OdysseyItemModels.GOLDEN_SPEAR, 2.5, 0.85, WeaponTypes.SPEAR)
 
-object StoneSpear: OdysseyWeapon("Stone Spear", Material.STONE_SHOVEL, OdysseyItemModels.STONE_SPEAR, 3.75, 0.85, WeaponTypes.POLE_ARM)
+object StoneSpear: OdysseyWeapon("Stone Spear", Material.STONE_SHOVEL, OdysseyItemModels.STONE_SPEAR, 3.75, 0.85, WeaponTypes.SPEAR)
 
-object IronSpear: OdysseyWeapon("Iron Spear", Material.IRON_SHOVEL, OdysseyItemModels.IRON_SPEAR, 4.75, 0.85, WeaponTypes.POLE_ARM)
+object IronSpear: OdysseyWeapon("Iron Spear", Material.IRON_SHOVEL, OdysseyItemModels.IRON_SPEAR, 4.75, 0.85, WeaponTypes.SPEAR)
 
-object DiamondSpear: OdysseyWeapon("Diamond Spear", Material.DIAMOND_SHOVEL, OdysseyItemModels.DIAMOND_SPEAR, 5.75, 0.85, WeaponTypes.POLE_ARM)
+object DiamondSpear: OdysseyWeapon("Diamond Spear", Material.DIAMOND_SHOVEL, OdysseyItemModels.DIAMOND_SPEAR, 5.75, 0.85, WeaponTypes.SPEAR)
 
-object NetheriteSpear: OdysseyWeapon("Netherite Spear", Material.NETHERITE_SHOVEL, OdysseyItemModels.NETHERITE_SPEAR, 6.75, 0.85, WeaponTypes.POLE_ARM)
+object NetheriteSpear: OdysseyWeapon("Netherite Spear", Material.NETHERITE_SHOVEL, OdysseyItemModels.NETHERITE_SPEAR, 6.75, 0.85, WeaponTypes.SPEAR)
 
 
 /* --------------------------HALBERD---------------------------*/
 
-object WoodenHalberd: OdysseyWeapon("Wooden Halberd", Material.WOODEN_SHOVEL, OdysseyItemModels.WOODEN_HALBERD, 4.5, 0.75, WeaponTypes.POLE_ARM)
+object WoodenHalberd: OdysseyWeapon("Wooden Halberd", Material.WOODEN_SHOVEL, OdysseyItemModels.WOODEN_HALBERD, 4.5, 0.75, WeaponTypes.SPEAR)
 
-object GoldenHalberd: OdysseyWeapon("Golden Halberd", Material.GOLDEN_SHOVEL, OdysseyItemModels.GOLDEN_HALBERD, 4.5, 0.8, WeaponTypes.POLE_ARM)
+object GoldenHalberd: OdysseyWeapon("Golden Halberd", Material.GOLDEN_SHOVEL, OdysseyItemModels.GOLDEN_HALBERD, 4.5, 0.8, WeaponTypes.SPEAR)
 
-object StoneHalberd: OdysseyWeapon("Stone Halberd", Material.STONE_SHOVEL, OdysseyItemModels.STONE_HALBERD, 5.5, 0.8, WeaponTypes.POLE_ARM)
+object StoneHalberd: OdysseyWeapon("Stone Halberd", Material.STONE_SHOVEL, OdysseyItemModels.STONE_HALBERD, 5.5, 0.8, WeaponTypes.SPEAR)
 
-object IronHalberd: OdysseyWeapon("Iron Halberd", Material.IRON_SHOVEL, OdysseyItemModels.IRON_HALBERD, 6.5, 0.8, WeaponTypes.POLE_ARM)
+object IronHalberd: OdysseyWeapon("Iron Halberd", Material.IRON_SHOVEL, OdysseyItemModels.IRON_HALBERD, 6.5, 0.8, WeaponTypes.SPEAR)
 
-object DiamondHalberd: OdysseyWeapon("Diamond Halberd", Material.DIAMOND_SHOVEL, OdysseyItemModels.DIAMOND_HALBERD, 7.5, 0.8, WeaponTypes.POLE_ARM)
+object DiamondHalberd: OdysseyWeapon("Diamond Halberd", Material.DIAMOND_SHOVEL, OdysseyItemModels.DIAMOND_HALBERD, 7.5, 0.8, WeaponTypes.SPEAR)
